requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
blinker==1.7.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.7
cssselect2==0.7.0
defusedxml==0.7.1
Flask==3.0.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.4
numpy==1.24.2
onfido==5.0.0
onfido-python==2.10.1
openai==0.27.4
pandas==1.5.3
pillow==10.2.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tinycss2==1.2.1
tqdm==4.65.0
twilio==8.10.3
urllib3==1.26.14
webencodings==0.5.1
Werkzeug==3.0.1
yarl==1.8.2
